+++ lua/codecompanion/strategies/chat/tools/files.lua
@@ -5,9 +5,6 @@ local log = require("codecompanion.utils.log")
 local util = require("codecompanion.utils.util")
 local xml2lua = require("codecompanion.utils.xml.xml2lua")
 
-local accepted = {}
-local rejected = {}
-
 ---Create a file and it's surrounding folders
 ---@param action table The action object
 ---@return nil
@@ -16,6 +13,15 @@ local function create(action)
   p.filename = p:expand()
   p:touch({ parents = true })
   p:write(action.contents or "", "w")
+end
+
+---Read the contents of af ile
+---@param action table The action object
+---@return string
+local function read(action)
+  local p = Path:new(action.path)
+  p.filename = p:expand()
+  return p:read()
 end
 
 ---Edit the contents of a file
@@ -78,6 +84,7 @@ end
 
 local actions = {
   create = create,
+  read = read,
   edit = edit,
   delete = delete,
   rename = rename,
@@ -85,36 +92,9 @@ local actions = {
   move = move,
 }
 
----Ask the user to approve the action
----@param action table The action object
-local function approve(action)
-  log:info("[Files Tool] Prompting for %s", action._attr.type)
-
-  local msg = string.upper(action._attr.type) .. " the file at " .. action.path
-  if action.new_path then
-    msg = msg .. " to " .. action.new_path
-  end
-  msg = msg .. "?"
-
-  local ok, choice = pcall(vim.fn.confirm, msg, "No\nYes")
-  if not ok or choice ~= 2 then
-    log:info("[Files Tool] Rejected the %s action", action._attr.type)
-    table.insert(rejected, {
-      type = action._attr.type,
-      path = action.path,
-      new_path = action.new_path,
-    })
-    return
-  end
-
-  log:info("[Files Tool] Approved the %s action", action._attr.type)
-  table.insert(accepted, {
-    type = action._attr.type,
-    path = action.path,
-    new_path = action.new_path,
-  })
-  return actions[action._attr.type](action)
-end
+---@class FilesTool.Output
+---@field status string The output status. Either "success" or "error"
+---@field msg string The message to send back to the LLM
 
 ---@class CodeCompanion.Tool
 return {
@@ -124,59 +104,25 @@ return {
     ---Execute the file commands
     ---@param self CodeCompanion.Tools The Tools object
     ---@param input any The output from the previous function call
-    ---@return table { status: string, output: string }
+    ---@return FilesTool.Output
     function(self, input)
-      accepted = {}
-      rejected = {}
-
-      local has_opts = config.strategies.agent.tools.files.opts
-
-      -- Run the action
-      local function run(action)
-        if has_opts and has_opts.user_approval then
-          approve(action)
-        else
-          actions[action._attr.type](action)
-          log:info("[Files Tool] Auto-approved the %s action", action._attr.type)
-        end
-      end
-
       -- Loop through the actions
       local action = self.tool.request.action
       if util.is_array(action) then
         for _, v in ipairs(action) do
-          run(v)
+          local ok, data = pcall(actions[action._attr.type], v)
+          if not ok then
+            return { status = "error", msg = data }
+          end
         end
       else
-        run(action)
+        local ok, data = pcall(actions[action._attr.type], action)
+        if not ok then
+          return { status = "error", msg = data }
+        end
       end
 
-      local function update_llm(tbl, str)
-        vim.iter(tbl):each(function(v)
-          local content = str .. string.upper(v.type) .. " the file at " .. v.path
-          if v.new_path then
-            content = content .. " to " .. v.new_path
-          end
-          self.chat:append_to_buf({
-            content = content .. "\n",
-          })
-        end)
-      end
-
-      if #accepted > 0 or #rejected > 0 then
-        self.chat:append_to_buf({
-          content = "Below is a list of all the file system actions that I accepted or rejected:\n\n",
-        })
-      end
-
-      if #accepted > 0 then
-        update_llm(accepted, "[x] ")
-      end
-      if #rejected > 0 then
-        update_llm(rejected, "[ ] ")
-      end
-
-      return { status = "success", output = nil }
+      return { status = "success", msg = nil }
     end,
   },
   schema = {
@@ -187,6 +133,15 @@ return {
           _attr = { type = "create" },
           path = "/Users/Oli/Code/new_app/hello_world.py",
           contents = "<![CDATA[    print('Hello World')]]>",
+        },
+      },
+    },
+    {
+      tool = {
+        _attr = { name = "files" },
+        action = {
+          _attr = { type = "read" },
+          path = "/Users/Oli/Code/new_app/hello_world.py",
         },
       },
     },
@@ -270,6 +225,7 @@ return {
   - **Don't escape** special characters
   - **Wrap contents in a CDATA block**, the contents could contain characters reserved by XML
   - **Don't duplicate code** in the response. Consider writing code directly into the contents tag of the XML
+  - The user's current working directory in Neovim is `%s`. They may refer to this in their message to you.
 
 4. **Actions**:
 
@@ -279,29 +235,37 @@ a) Create:
 %s
 ```
 - This will ensure a file is created at the specified path with the given content.
+- It will also ensure that any folders that don't exist in the path are created.
 
-b) Edit:
+b) Read:
+
+```xml
+%s
+```
+- This will output the contents of a file at the specified path.
+
+c) Edit:
 
 ```xml
 %s
 ```
 - This will ensure a file is edited at the specified path and its contents replaced with the given content.
 
-c) Delete:
+d) Delete:
 
 ```xml
 %s
 ```
 - This will ensure a file is deleted at the specified path.
 
-d) Rename:
+e) Rename:
 
 ```xml
 %s
 ```
 - Ensure `new_path` contains the filename
 
-e) Copy:
+f) Copy:
 
 ```xml
 %s
@@ -309,7 +273,7 @@ e) Copy:
 - Ensure `new_path` contains the filename
 - Any folders that don't exist in the path will be created
 
-f) Move:
+g) Move:
 
 ```xml
 %s
@@ -326,12 +290,14 @@ f) Move:
 Remember:
 - Minimize explanations unless prompted. Focus on generating correct XML.
 - If the user types `~` in their response, do not replace or expand it.]],
+      vim.fn.getcwd(),
       xml2lua.toXml({ tools = { schema[1] } }),
       xml2lua.toXml({ tools = { schema[2] } }),
       xml2lua.toXml({ tools = { schema[3] } }),
       xml2lua.toXml({ tools = { schema[4] } }),
       xml2lua.toXml({ tools = { schema[5] } }),
       xml2lua.toXml({ tools = { schema[6] } }),
+      xml2lua.toXml({ tools = { schema[7] } }),
       xml2lua.toXml({
         tools = {
           tool = {
@@ -345,4 +311,58 @@ Remember:
       })
     )
   end,
+  handlers = {
+    ---Approve the command to be run
+    ---@param self CodeCompanion.Tools The tool object
+    ---@param action table
+    ---@return boolean
+    approved = function(self, action)
+      log:info("[Files Tool] Prompting for %s", action._attr.type)
+
+      local msg = string.upper(action._attr.type) .. " the file at " .. action.path
+      if action.new_path then
+        msg = msg .. " to " .. action.new_path
+      end
+      msg = msg .. "?"
+
+      local ok, choice = pcall(vim.fn.confirm, msg, "No\nYes")
+      if not ok or choice ~= 2 then
+        log:info("[Files Tool] Rejected the %s action", action._attr.type)
+        return false
+      end
+
+      log:info("[Files Tool] Approved the %s action", action._attr.type)
+      return true
+    end,
+  },
+  output = {
+    success = function(self, action, output)
+      return self.chat:add_buf_message({
+        role = config.constants.USER_ROLE,
+        content = string.format("The %s action was executed successfully.\n\n", string.upper(action._attr.type)),
+      })
+    end,
+
+    error = function(self, action, err)
+      return self.chat:add_buf_message({
+        role = config.constants.USER_ROLE,
+        content = string.format(
+          [[There was an error running the %s action:
+
+```txt
+%s
+```]],
+          string.upper(action._attr.type),
+          err
+        ),
+      })
+    end,
+
+    rejected = function(self, action)
+      return self.chat:add_buf_message({
+        role = config.constants.USER_ROLE,
+        content = string.format("I rejected the %s action.\n\n", string.upper(action._attr.type)),
+      })
+    end,
+  },
 }